// project/src/types/training-time.enum.ts
export enum TrainingTime {
  TenToThirty = '10-30 мин',
  ThirtyToFifty= '30-50 мин',
  FiftyToEighty = '50-80 мин',
  OverEighty = 'больше 80 мин',
}

export enum TrainingTimeEn {
  TenToThirty = 'TenToThirty',
  ThirtyToFifty= 'ThirtyToFifty',
  FiftyToEighty = 'FiftyToEighty',
  OverEighty = 'OverEighty',
}