// project/src/types/training-type.enum.ts
export enum TrainingType {
  Yoga = 'йога',
  Running= 'бег',
  Boxing = 'бокс',
  Stretching = 'стрейчинг',
  Crossfit = 'кроссфит',
  Aerobics = 'аэробика',
  Pilates = 'пилатес'
}

export enum TrainingTypeEn {
  Yoga = 'Yoga',
  Running= 'Running',
  Boxing = 'Boxing',
  Stretching = 'Stretching',
  Crossfit = 'Crossfit',
  Aerobics = 'Aerobics',
  Pilates = 'Pilates'
}